+++ programmable_video/android/src/main/kotlin/twilio/flutter/twilio_programmable_video/PluginHandler.kt
@@ -13,6 +13,8 @@ import android.os.Handler
 import android.os.Looper
 import androidx.annotation.NonNull
 import com.twilio.video.AudioCodec
+import com.twilio.video.Camera2Capturer
+import com.twilio.video.CameraCapturer
 import com.twilio.video.ConnectOptions
 import com.twilio.video.DataTrackOptions
 import com.twilio.video.G722Codec
@@ -111,19 +113,31 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
             "RemoteAudioTrack#enablePlayback" -> remoteAudioTrackEnable(call, result)
             "RemoteAudioTrack#isPlaybackEnabled" -> isRemoteAudioTrackPlaybackEnabled(call, result)
             "CameraCapturer#switchCamera" -> switchCamera(call, result)
-            "CameraCapturer#hasTorch" -> hasTorch(result)
             "CameraCapturer#setTorch" -> setTorch(call, result)
+            "CameraSource#getSources" -> getSources(call, result)
             else -> result.notImplemented()
         }
     }
 
-    private fun switchCamera(call: MethodCall, result: MethodChannel.Result) {
-        TwilioProgrammableVideoPlugin.debug("PluginHandler::switchCamera => called")
-        VideoCapturerHandler.switchCamera(call, result)
+    private fun getSources(call: MethodCall, result: MethodChannel.Result) {
+        TwilioProgrammableVideoPlugin.debug("PluginHandler.getSources => called")
+        return result.success(TwilioProgrammableVideoPlugin.cameraEnumerator.deviceNames.map {
+            VideoCapturerHandler.cameraIdToMap(it)
+        })
     }
 
-    private fun hasTorch(result: MethodChannel.Result) {
-        VideoCapturerHandler.hasTorch(result)
+    private fun switchCamera(call: MethodCall, result: MethodChannel.Result) {
+        TwilioProgrammableVideoPlugin.debug("PluginHandler.switchCamera => called")
+        val newCameraId = call.argument<String>("cameraId")
+                ?: return result.error("MISSING_PARAMS", "The parameter 'cameraId' was not given", null)
+
+        val capturer = TwilioProgrammableVideoPlugin.cameraCapturer
+        if (capturer is Camera2Capturer)
+            capturer.switchCamera(newCameraId)
+        else if (capturer is CameraCapturer)
+            capturer.switchCamera(newCameraId)
+
+        return result.success(VideoCapturerHandler.videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer!!, newCameraId))
     }
 
     private fun setTorch(call: MethodCall, result: MethodChannel.Result) {
@@ -216,12 +230,12 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
     }
 
     private fun getRemoteAudioTrack(sid: String): RemoteAudioTrackPublication? {
-        val remoteParticipants = TwilioProgrammableVideoPlugin.roomListener?.room?.remoteParticipants
+        val remoteParticipants = TwilioProgrammableVideoPlugin.roomListener.room?.remoteParticipants
             ?: return null
 
-        var remoteAudioTrack: RemoteAudioTrackPublication? = null
+        var remoteAudioTrack: RemoteAudioTrackPublication?
         for (remoteParticipant in remoteParticipants) {
-            remoteAudioTrack = remoteParticipant.remoteAudioTracks.firstOrNull { it.trackSid.equals(sid) }
+            remoteAudioTrack = remoteParticipant.remoteAudioTracks.firstOrNull { it.trackSid == sid }
             if (remoteAudioTrack != null) return remoteAudioTrack
         }
         return null
@@ -490,14 +504,14 @@ class PluginHandler : MethodCallHandler, ActivityAware, BaseListener {
                     videoTrack as Map<*, *> // Ensure right type.
                     val videoCapturerMap = videoTrack["videoCapturer"] as Map<*, *>
 
-                    if ((videoCapturerMap["type"] as String).equals("CameraCapturer")) {
+                    if ((videoCapturerMap["type"] as String) == "CameraCapturer") {
                         VideoCapturerHandler.initializeCapturer(videoCapturerMap, result)
                     } else {
                         return result.error("INIT_ERROR", "VideoCapturer type ${videoCapturerMap["type"]} not yet supported.", null)
                     }
 
                     if (TwilioProgrammableVideoPlugin.cameraCapturer != null) {
-                        videoTracks.add(LocalVideoTrack.create(this.applicationContext, videoTrack["enable"] as Boolean, TwilioProgrammableVideoPlugin.cameraCapturer, videoTrack["name"] as String))
+                        videoTracks.add(LocalVideoTrack.create(this.applicationContext, videoTrack["enable"] as Boolean, TwilioProgrammableVideoPlugin.cameraCapturer!!, videoTrack["name"] as String))
                     }
                 }
                 TwilioProgrammableVideoPlugin.debug("PluginHandler::connect => setting videoTracks to '${videoTracks.joinToString(", ")}'")