+++ programmable_video/lib/src/camera_capturer.dart
@@ -23,10 +23,10 @@ class CameraCapturer implements VideoCapturer {
   /// Called when the camera has thrown an error
   Stream<CameraErrorEvent>? onCameraError;
 
-  CameraSource? _cameraSource;
+  CameraSource? _source;
 
   /// The current specified camera source.
-  CameraSource? get cameraSource => _cameraSource;
+  CameraSource? get source => _source;
 
   /// Indicates that the camera capturer is not a screen cast.
   @override
@@ -35,9 +35,9 @@ class CameraCapturer implements VideoCapturer {
   /// Singleton factory.
   ///
   /// Only one instance is allowed.
-  factory CameraCapturer(CameraSource? cameraSource) {
+  factory CameraCapturer(CameraSource? source) {
     _cameraCapturer ??= CameraCapturer._internal();
-    _cameraCapturer!._cameraSource = cameraSource;
+    _cameraCapturer!._source = source;
     _cameraCapturer!._cameraStream ??= ProgrammableVideoPlatform.instance.cameraStream()?.listen(_cameraCapturer!._parseCameraEvents);
     _cameraCapturer!.onCameraSwitched ??= _cameraCapturer!._onCameraSwitched.stream;
     _cameraCapturer!.onFirstFrameAvailable ??= _cameraCapturer!._onFirstFrameAvailable.stream;
@@ -73,16 +73,16 @@ class CameraCapturer implements VideoCapturer {
     onCameraError = null;
   }
 
-  /// Switch the current [CameraSource].
+  /// Switch to the given [CameraSource].
   ///
   /// This method can be invoked while capturing frames or not.
-  /// Throws a [FormatException] if the result could not be parsed to a [CameraSource].
-  /// Throws a [MissingCameraException] if no camera is found for the [CameraSource]
-  /// that is not presently in use.
+  /// Throws a [MissingCameraException] if no camera is found for the [CameraSource].
   /// Throws a [NotFoundException] when the [CameraCapturer] was not provided at time of connection.
-  Future<void> switchCamera() async {
+  Future<void> switchCamera(CameraSource newSource) async {
     try {
-      _cameraSource = await ProgrammableVideoPlatform.instance.switchCamera();
+      _source = await ProgrammableVideoPlatform.instance.switchCamera(
+        newSource,
+      );
     } on PlatformException catch (err) {
       throw TwilioProgrammableVideo._convertException(err);
     }
@@ -90,11 +90,8 @@ class CameraCapturer implements VideoCapturer {
 
   /// Get availability of torch on active [CameraSource].
   ///
-  /// This method can be invoked while capturing frames or not.
-  /// Returns false if there is no active camera.
-  Future<bool> hasTorch() async {
-    return ProgrammableVideoPlatform.instance.hasTorch();
-  }
+  /// Returns false if there is no active [CameraSource].
+  bool get hasTorch => source?.hasTorch ?? false;
 
   /// Set state of torch on active [CameraSource].
   ///
@@ -117,7 +114,7 @@ class CameraCapturer implements VideoCapturer {
   @override
   void _updateFromModel(VideoCapturerModel? model) {
     if (model is CameraCapturerModel && model.source != null) {
-      _cameraSource = model.source!;
+      _source = model.source;
     }
   }