+++ programmable_video/lib/src/parts.dart
@@ -14,6 +14,8 @@ export 'package:twilio_programmable_video_platform_interface/src/audio_codecs/au
 export 'package:twilio_programmable_video_platform_interface/src/enums/enum_exports.dart';
 export 'package:twilio_programmable_video_platform_interface/src/video_codecs/video_codec.dart';
 
+export 'package:twilio_programmable_video_platform_interface/src/camera_source.dart';
+
 part 'audio_track.dart';
 part 'audio_track_publication.dart';
 part 'camera_capturer.dart';