+++ programmable_video_platform_interface/lib/twilio_programmable_video_platform_interface.dart
@@ -3,3 +3,4 @@
 library twilio_programmable_video_platform_interface;
 
 export 'src/programmable_video_platform_interface.dart';
+export 'src/camera_source.dart';