+++ programmable_video/CHANGELOG.md
@@ -1,6 +1,22 @@
+## 0.10.0
+
+- **BREAKING**: Migrated TwilioVideo iOS SDK from v3 to v4.
+- **BREAKING**: Migrated TwilioVideo Android SDK from v5 to v6.
+- **BREAKING**: Replaced `CameraSource` enum with a class variant to represent a potential source for camera capturing.
+- **BREAKING**: The `CameraCapturer.hasTorch()` method has been replaced with a simple getter to it's `CameraSource` `hasTorch` property.
+- Retrieving `CameraSource`s can be done using `CameraSource.getSources()` method.
+
+## 0.9.0+2
+
+- Fixed remote video stats (replaced an incorrect map key).
+
+## 0.9.0+1
+
+- Fixed typo in stats (trackSide -> trackSid).
+
 ## 0.9.0
 
-- **BREAKING** Made `exception` property of ConnectFailure, Reconnecting and Disconnected events nullable.
+- **BREAKING**: Made `exception` property of ConnectFailure, Reconnecting and Disconnected events nullable.
 
 ## 0.8.0+1